Create Disclaimer.txt
This tool may be used for legal purposes only.  Users take full responsibility
for any actions performed using this tool.  The author accepts no liability for
damage caused by this tool.  If these terms are not acceptable to you, then do 
not use this tool.

In all other respects the GPL version 2 applies.